GitHub specific helpers

installGitHub is a vectorized installGithubPackages. This will attempt to identify all dependencies of all supplied packages first, then load the packages in the correct order so that each of their dependencies are met before each is installed.

parseGitHub(pkgDT, verbose = getOption("Require.verbose"))

installGitHub(
  pkgDT,
  toInstall,
  install_githubArgs = list(),
  dots = dots,
  verbose = getOption("Require.verbose")
)

Arguments

pkgDT

A character string with full package names or a data.table with at least 2 columns "Package" and "packageFullName".

verbose

Numeric or logical indicating how verbose should the function be. If -1 or less, then as little verbosity as possible. If 0 or FALSE, then minimal outputs; if 1 or TRUE, more outputs; 2 even more. NOTE: in Require function, when verbose >= 2, the return object will have an attribute: attr(.., "Require") which has lots of information about the processes of the installs.

toInstall

DESCRIPTION NEEDED

install_githubArgs

Any arguments passed to install_github

dots

A list of ..., e.g., list(...). Only for internal use.

Value

parseGitHub returns a data.table with added columns.

installGitHub returns a named character vector indicating packages successfully installed, unless the word "Failed" is returned, indicating installation failure. The names will be the full GitHub package name, as provided to gitPkgNames in the function call.

Details

parseGitHub turns the single character string representation into 3 or 4: Account, Repo, Branch, SubFolder.

Examples

if (FALSE) {
  installGitHub(c("PredictiveEcology/Require", "PredictiveEcology/quickPlot"))
}