Internals used by Require

While these are not intended to be called manually by users, they may be of some use for advanced users.

getPkgVersions(pkgDT, install = TRUE, verbose = getOption("Require.verbose"))

getAvailable(
  pkgDT,
  purge = FALSE,
  repos = getOption("repos"),
  verbose = getOption("Require.verbose")
)

installFrom(
  pkgDT,
  purge = FALSE,
  repos = getOption("repos"),
  verbose = getOption("Require.verbose")
)

doInstalls(
  pkgDT,
  install_githubArgs,
  install.packagesArgs,
  install = TRUE,
  repos = getOption("repos"),
  verbose = getOption("Require.verbose"),
  ...
)

doLoading(pkgDT, require = TRUE, verbose = getOption("Require.verbose"), ...)

archiveVersionsAvailable(package, repos)

Arguments

pkgDT

A character string with full package names or a data.table with at least 2 columns "Package" and "packageFullName".

install

Logical or "force". If FALSE, this will not try to install anything. If "force", then it will force installation of requested packages, mimicking a call to e.g., install.packages. If TRUE, the default, then this function will try to install any missing packages or dependencies.

verbose

Numeric or logical indicating how verbose should the function be. If -1 or less, then as little verbosity as possible. If 0 or FALSE, then minimal outputs; if 1 or TRUE, more outputs; 2 even more. NOTE: in Require function, when verbose >= 2, the return object will have an attribute: attr(.., "Require") which has lots of information about the processes of the installs.

purge

Logical. Should all caches be purged? Default is getOption("Require.purge", FALSE). There is a lot of internal caching of results throughout the Require package. These help with speed and reduce calls to internet sources. However, sometimes these caches must be purged. The cached values are renewed when found to be too old, with the age limit. This maximum age can be set in seconds with the environment variable R_AVAILABLE_PACKAGES_CACHE_CONTROL_MAX_AGE, or if unset, defaults to 3600 (one hour -- see utils::available.packages).

Internally, there are calls to available.packages.

repos

The remote repository (e.g., a CRAN mirror), passed to either install.packages, install_github or installVersions.

install_githubArgs

List of optional named arguments, passed to install.packages inside installGitHubPackage.

install.packagesArgs

List of optional named arguments, passed to install.packages.

...

Passed to install.packages. Good candidates are e.g., type or dependencies. This can be used with install_githubArgs or install.packageArgs which give individual options for those 2 internal function calls.

require

Logical or character string. If TRUE, the default, then the function will attempt to call require on all requested packages, possibly after they are installed. If a character string, then it will only call require on those specific packages (i.e., it will install the ones listed in packages, but load the packages listed in require)

package

A single package name (without version or github specifications)

Value

In general, these functions return a data.table with various package information, installation status, version, available version etc.

Details

doInstall is a wrapper around utils::install.packages, installGithub, and installCRAN, and installArchive

doLoading is a wrapper around require.

archiveVersionsAvailable searches CRAN Archives for available versions. It has been borrowed from a sub-set of the code in a non-exported function: remotes:::download_version_url